+++ BUILD.bazel
@@ -2725,13 +2725,3 @@ genrule(
     """,
     local = 1,
 )
-
-ray_cc_library(
-    name = "cgroup_setup",
-    srcs = ["src/ray/raylet/cgroup_setup.cc"],
-    hdrs = ["src/ray/raylet/cgroup_setup.h"],
-    deps = [
-        "//src/ray/util",
-        "@com_google_absl//absl/strings:str_format",
-    ],
-)
