+++ src/ray/common/cgroup/cgroup_setup.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "src/ray/raylet/cgroup_setup.h"
+#include "src/ray/common/cgroup/cgroup_setup.h"
 
 #include <sys/stat.h>
